Parse Old Bailey Trials

For parsing Old Bailey trial data.

Usage

parse_trials(
  xml_address
  )

Arguments

xml_address

One or more XML addresses. A single XML address can be passed as a string, or numerous XML addresses can be passed as a list.

Value

Dataframe containing API pull of parsed Old Bailey Trials.

Examples

# Return a dataframe with parsed trial data.
xml_address <- "https://www.oldbaileyonline.org/obapi/text?div=t17690112-9"
parsed_trial <- parse_trials(xml_address)
#> [1] "Parsing https://www.oldbaileyonline.org/obapi/text?div=t17690112-9"
#> [1] "1 out of 1 trials."
head(parsed_trial)
#>   trial_account_id  defendant_name defendant_gender
#> 1      t17690112-9 Matthew Skinner             male
#> 2      t17690112-9 Matthew Skinner             male
#> 3      t17690112-9 Matthew Skinner             male
#> 4      t17690112-9 Matthew Skinner             male
#> 5      t17690112-9 Matthew Skinner             male
#> 6      t17690112-9 Matthew Skinner             male
#>                          victim_name victim_gender        crime_location
#> 1 William Saunders, Benjamin Hopkins    male, male Aldgate, Cripple-gate
#> 2 William Saunders, Benjamin Hopkins    male, male Aldgate, Cripple-gate
#> 3 William Saunders, Benjamin Hopkins    male, male Aldgate, Cripple-gate
#> 4 William Saunders, Benjamin Hopkins    male, male Aldgate, Cripple-gate
#> 5 William Saunders, Benjamin Hopkins    male, male Aldgate, Cripple-gate
#> 6 William Saunders, Benjamin Hopkins    male, male Aldgate, Cripple-gate
#>       offence_category offence_subcategory punishment_category
#> 1 deception, deception    forgery, forgery                    
#> 2 deception, deception    forgery, forgery                    
#> 3 deception, deception    forgery, forgery                    
#> 4 deception, deception    forgery, forgery                    
#> 5 deception, deception    forgery, forgery                    
#> 6 deception, deception    forgery, forgery                    
#>   punishment_subcategory              verdict verdict_subcategory speech_id
#> 1                        notGuilty, notGuilty                             0
#> 2                        notGuilty, notGuilty                             1
#> 3                        notGuilty, notGuilty                             2
#> 4                        notGuilty, notGuilty                             3
#> 5                        notGuilty, notGuilty                             4
#> 6                        notGuilty, notGuilty                             5
#>       speaker_name
#> 1                 
#> 2 William Saunders
#> 3               Q.
#> 4          Vaughn.
#> 5                 
#> 6               Q.
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            body_text
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        90. (L.) Matthew Skinner was indicted for  forging, and counteirfeting an order, for the payment of money, directed to Mess. Fuller, Baker, and Halford , to this purport: "October "2 l, 1768. Pay to Mr. James Brown , or "bearer, 15 l. 15 s. and 6 d. G. Humphrys." And for publishing the same, with intent to defraud William Saunders , October 22 .
#> 2 On the 22d of October last, the prisoner came to my shop; I am a hosier  near Aldgate ; he told me, he was recommended to buy three or four dozen of house, of 23 or 24 s. a dozen, and that they were for a person at Berry, who did a great deal of business in my way, and if we used him well, he would be a good customer; he looked out five dozen, which came to 5 l. 6 s. I asked him, what name I should make the bill of parcels in; he said, his name was Robinson; he then presented a draft for 15 l. 15 s. and 6 d. and desired me to give him change. He said, it was the last draft he had. I was hesitating upon it, but he appearing to be a reputable country shop-keeper, I was inclinable to give him change, with this precaution, that as soon as I had given him change, which was a draft upon Sir Richard Glyn and Co. I intended to go immediately to know whether it was a good one. I gave him a draft upon Sir Richard for 10 l. 9 s. and 6 d. which with the 5 l. 6 s. made the money. I went to Mr. Fuller's, and found his bill was a bad one. Sir Richard Glyn 's house being but two doors from that, I got to Sir Richard Glyn 's four minutes before the prisoner came in. Skinner seeing me, he wanted to withdraw. He said, he would call again; they told him, he had better do his business now. He had my draft in his hand, and laid it on the counter. We went for a constable, and took him in custody. He said, he received it of one Gompy, and said, his own name was Jonathan Robinson , that he came to town about selling a horse, and that he could neither write nor read. We went before an Alderman on the Monday. ( This was Saturday evening.) He there acknowledged another fraud of this kind. He said, he came from Sheffield, and that he received it of one Gompy, at a house in Goodman's Fields. He said, he should be glad to have Gompy taken. John Vaughn. The first time I saw the prisoner, was at Sir Richard Glyn 's. I went along with him and Mr. Saunders, and Mr. Halford, to a tavern in the Poultry. He was asked his name. He said, it was Jonathan Robinson , and was lately come out of the country, I think, from Sheffield, and had the draft of one Gompy, a Jew, and that he himself could neither write nor read. That he had laid in Goodman's-Fields the night before, and had taken it there; that he had been in town but a few days; that he lodged, when he came to town, at the Three Nuns Inn, in White-Chapel, three or four nights, but had laid at the Fleece Goodman's-fields the night before; and that he came to town to dispose of a horse. But he was not consistent in the account that he gave of himself.
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Does any body of the name of G. Humphrys keep cash at your house?
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         I have been there fifteen or sixteen years, and I know of no such person. The draft is a blank cheque of ours that has seen filled up (Produced in court.) It is read to the purport as in the indictment.
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     Prisoner's Defence. I received these drafts of Gompy, in order to get them off for him. I was very innocent of any bad design.
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             to Saunders. How came you to find that the prisoner's name was not Robinson? Saunders. When he was before Mr. Alderman Kennet, at the Mansion-House, he said his name was Matthew Skinner . The Alderman asked him, why he had said his name was Jonathan Robinson ; he answered, he was then confused, and did not know what he said.
#>       date                                                xml_address
#> 1 17690112 https://www.oldbaileyonline.org/obapi/text?div=t17690112-9
#> 2 17690112 https://www.oldbaileyonline.org/obapi/text?div=t17690112-9
#> 3 17690112 https://www.oldbaileyonline.org/obapi/text?div=t17690112-9
#> 4 17690112 https://www.oldbaileyonline.org/obapi/text?div=t17690112-9
#> 5 17690112 https://www.oldbaileyonline.org/obapi/text?div=t17690112-9
#> 6 17690112 https://www.oldbaileyonline.org/obapi/text?div=t17690112-9